fix bugs
save ROIs.
there are several save functions, for each one, when use them, should
not affact other data. For example, before fixing the bug, when save
ROIs, it will affact the notes data.So be careful when writing these
functions, each save function only change some data in the metadata but
not others.