fixed bugs to save stimulus, notes
The save functions in stim, notewriter class use pwd to find the path to
save. However,  the path found in this way was not consistent with the
path of opened file ( not figured out why yet). So I used another way to
solve this problem. When open an image file, I put the filedir into the
imge figure UserData, then in the save functions of these two classes,
use findobj function to get the handle of the image figure, then exact
the filedir from the UserDate.